private access to registered users

+++ controllers/contactControllers.js
@@ -13,7 +13,8 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
 const getContact = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find({ user_id: req.user.id });
+  // First we need to fetch all contacts of a logged in user so there it is: user_id: req.user.id
+  const contacts = await Contact.find({ id: req.user.id });
   res.status(200).json(contacts);
 });
 
@@ -27,7 +28,6 @@ const getContactID = asyncHandler(async (req, res) => {
 });
 
 const postContact = asyncHandler(async (req, res) => {
-  console.log(req.body);
   const { name, email, phone } = req.body;
   if (!name || !email || !phone) {
     res.status(400);
@@ -38,6 +38,7 @@ const postContact = asyncHandler(async (req, res) => {
     name,
     email,
     phone,
+    id: req.user.id,
   });
   res.json(contact);
 });
@@ -48,11 +49,18 @@ const putContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found!");
   }
+// To check if a diff user is trying to update the contact  
+if(contacts.id.toString() !== req.user.id)
+{
+    res.status(403);
+    throw new Error("User dont have permission to update other users contact");
+  }
+
   const updatedContact = await Contact.findByIdAndUpdate(
     req.params.id,
     req.body,
     { new: true }
-  );  
+  );
   res.json(updatedContact);
 });
 
@@ -62,7 +70,11 @@ const deleteContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found!");
   }
-  const deletedContact = await Contact.deleteOne({ _id: req.params.id })
+   if (contacts.id.toString() !== req.user.id) {
+     res.status(403);
+     throw new Error("User dont have permission to delete other users contact");
+   }
+  const deletedContact = await Contact.deleteOne({ _id: req.params.id });
   res.json({ deletedContact });
 });
